add option to disable PlanBuild patch to allow future fixes from PlanBuild
