--- ValheimRAFT.Util/ValheimRAFTServer.cs
@@ -1,13 +1,0 @@
-﻿namespace ValheimRAFT.Util;
-
-public class ValheimRAFTServer
-{
-  public string hello = "1";
-
-  public string isServer { get; set; }
-
-  public ValheimRAFTServer()
-  {
-    // ValheimRAFTServer.isServer = ZN
-  }
-}
